@@ -1,9 +1,9 @@
 // Code generated by MockGen. DO NOT EDIT.
-// Source: github.com/TimeWtr/Chanjet/core (interfaces: EvictPolicy)
+// Source: github.com/TimeWtr/TurboStream/core (interfaces: EvictPolicy)
 //
 // Generated by this command:
 //
-//	mockgen -destination=./mocks/queue/evictpolicy_mock.go -package queue_mocks github.com/TimeWtr/Chanjet/core EvictPolicy
+//	mockgen -destination=./mocks/queue/evictpolicy_mock.go -package queue_mocks github.com/TimeWtr/TurboStream/core EvictPolicy
 //
 
 // Package queue_mocks is a generated GoMock package.
@@ -12,7 +12,7 @@ package queue_mocks
 import (
 	reflect "reflect"
 
-	component "github.com/TimeWtr/Chanjet/core/component"
+	component "github.com/TimeWtr/TurboStream/core/component"
 	gomock "go.uber.org/mock/gomock"
 )
 
